[SIEM][Detection Engine][Lists] Moves getQueryFilter to common folder for use by both front and backend
## Summary

* Moves querying and tests from server to common

The function we are interested using on the front end is:

```ts
export const getQueryFilter = (
  query: Query,
  language: Language,
  filters: Array<Partial<Filter>>,
  index: Index,
  lists: ExceptionListItemSchema[]
) => {
```

### Checklist

- [x] [Unit or functional tests](https://github.com/elastic/kibana/blob/master/CONTRIBUTING.md#cross-browser-compatibility) were updated or added to match the most common scenarios
