+++ x-pack/plugins/security_solution/common/detection_engine/get_query_filter.test.ts
@@ -1,0 +1,650 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+import { getQueryFilter } from './get_query_filter';
+import { Filter } from 'src/plugins/data/public';
+import { getExceptionListItemSchemaMock } from '../../../lists/common/schemas/response/exception_list_item_schema.mock';
+
+describe('get_filter', () => {
+  describe('getQueryFilter', () => {
+    test('it should work with an empty filter as kuery', () => {
+      const esQuery = getQueryFilter('host.name: linux', 'kuery', [], ['auditbeat-*'], []);
+      expect(esQuery).toEqual({
+        bool: {
+          must: [],
+          filter: [
+            {
+              bool: {
+                should: [
+                  {
+                    match: {
+                      'host.name': 'linux',
+                    },
+                  },
+                ],
+                minimum_should_match: 1,
+              },
+            },
+          ],
+          should: [],
+          must_not: [],
+        },
+      });
+    });
+
+    test('it should work with an empty filter as lucene', () => {
+      const esQuery = getQueryFilter('host.name: linux', 'lucene', [], ['auditbeat-*'], []);
+      expect(esQuery).toEqual({
+        bool: {
+          must: [
+            {
+              query_string: {
+                query: 'host.name: linux',
+                analyze_wildcard: true,
+                time_zone: 'Zulu',
+              },
+            },
+          ],
+          filter: [],
+          should: [],
+          must_not: [],
+        },
+      });
+    });
+
+    test('it should work with a simple filter as a kuery', () => {
+      const esQuery = getQueryFilter(
+        'host.name: windows',
+        'kuery',
+        [
+          {
+            meta: {
+              alias: 'custom label here',
+              disabled: false,
+              key: 'host.name',
+              negate: false,
+              params: {
+                query: 'siem-windows',
+              },
+              type: 'phrase',
+            },
+            query: {
+              match_phrase: {
+                'host.name': 'siem-windows',
+              },
+            },
+          },
+        ],
+        ['auditbeat-*'],
+        []
+      );
+      expect(esQuery).toEqual({
+        bool: {
+          must: [],
+          filter: [
+            {
+              bool: {
+                should: [
+                  {
+                    match: {
+                      'host.name': 'windows',
+                    },
+                  },
+                ],
+                minimum_should_match: 1,
+              },
+            },
+            {
+              match_phrase: {
+                'host.name': 'siem-windows',
+              },
+            },
+          ],
+          should: [],
+          must_not: [],
+        },
+      });
+    });
+
+    test('it should work with a simple filter as a kuery without meta information', () => {
+      const esQuery = getQueryFilter(
+        'host.name: windows',
+        'kuery',
+        [
+          {
+            query: {
+              match_phrase: {
+                'host.name': 'siem-windows',
+              },
+            },
+          },
+        ],
+        ['auditbeat-*'],
+        []
+      );
+      expect(esQuery).toEqual({
+        bool: {
+          must: [],
+          filter: [
+            {
+              bool: {
+                should: [
+                  {
+                    match: {
+                      'host.name': 'windows',
+                    },
+                  },
+                ],
+                minimum_should_match: 1,
+              },
+            },
+            {
+              match_phrase: {
+                'host.name': 'siem-windows',
+              },
+            },
+          ],
+          should: [],
+          must_not: [],
+        },
+      });
+    });
+
+    test('it should work with a simple filter as a kuery without meta information with an exists', () => {
+      const query: Partial<Filter> = {
+        query: {
+          match_phrase: {
+            'host.name': 'siem-windows',
+          },
+        },
+      };
+
+      const exists: Partial<Filter> = {
+        exists: {
+          field: 'host.hostname',
+        },
+      } as Partial<Filter>;
+
+      const esQuery = getQueryFilter(
+        'host.name: windows',
+        'kuery',
+        [query, exists],
+        ['auditbeat-*'],
+        []
+      );
+      expect(esQuery).toEqual({
+        bool: {
+          must: [],
+          filter: [
+            {
+              bool: {
+                should: [
+                  {
+                    match: {
+                      'host.name': 'windows',
+                    },
+                  },
+                ],
+                minimum_should_match: 1,
+              },
+            },
+            {
+              match_phrase: {
+                'host.name': 'siem-windows',
+              },
+            },
+            {
+              exists: {
+                field: 'host.hostname',
+              },
+            },
+          ],
+          should: [],
+          must_not: [],
+        },
+      });
+    });
+
+    test('it should work with a simple filter that is disabled as a kuery', () => {
+      const esQuery = getQueryFilter(
+        'host.name: windows',
+        'kuery',
+        [
+          {
+            meta: {
+              alias: 'custom label here',
+              disabled: true,
+              key: 'host.name',
+              negate: false,
+              params: {
+                query: 'siem-windows',
+              },
+              type: 'phrase',
+            },
+            query: {
+              match_phrase: {
+                'host.name': 'siem-windows',
+              },
+            },
+          },
+        ],
+        ['auditbeat-*'],
+        []
+      );
+      expect(esQuery).toEqual({
+        bool: {
+          must: [],
+          filter: [
+            {
+              bool: {
+                should: [
+                  {
+                    match: {
+                      'host.name': 'windows',
+                    },
+                  },
+                ],
+                minimum_should_match: 1,
+              },
+            },
+          ],
+          should: [],
+          must_not: [],
+        },
+      });
+    });
+
+    test('it should work with a simple filter as a lucene', () => {
+      const esQuery = getQueryFilter(
+        'host.name: windows',
+        'lucene',
+        [
+          {
+            meta: {
+              alias: 'custom label here',
+              disabled: false,
+              key: 'host.name',
+              negate: false,
+              params: {
+                query: 'siem-windows',
+              },
+              type: 'phrase',
+            },
+            query: {
+              match_phrase: {
+                'host.name': 'siem-windows',
+              },
+            },
+          },
+        ],
+        ['auditbeat-*'],
+        []
+      );
+      expect(esQuery).toEqual({
+        bool: {
+          must: [
+            {
+              query_string: {
+                query: 'host.name: windows',
+                analyze_wildcard: true,
+                time_zone: 'Zulu',
+              },
+            },
+          ],
+          filter: [
+            {
+              match_phrase: {
+                'host.name': 'siem-windows',
+              },
+            },
+          ],
+          should: [],
+          must_not: [],
+        },
+      });
+    });
+
+    test('it should work with a simple filter that is disabled as a lucene', () => {
+      const esQuery = getQueryFilter(
+        'host.name: windows',
+        'lucene',
+        [
+          {
+            meta: {
+              alias: 'custom label here',
+              disabled: true,
+              key: 'host.name',
+              negate: false,
+              params: {
+                query: 'siem-windows',
+              },
+              type: 'phrase',
+            },
+            query: {
+              match_phrase: {
+                'host.name': 'siem-windows',
+              },
+            },
+          },
+        ],
+        ['auditbeat-*'],
+        []
+      );
+      expect(esQuery).toEqual({
+        bool: {
+          must: [
+            {
+              query_string: {
+                query: 'host.name: windows',
+                analyze_wildcard: true,
+                time_zone: 'Zulu',
+              },
+            },
+          ],
+          filter: [],
+          should: [],
+          must_not: [],
+        },
+      });
+    });
+
+    test('it should work with a list', () => {
+      const esQuery = getQueryFilter(
+        'host.name: linux',
+        'kuery',
+        [],
+        ['auditbeat-*'],
+        [getExceptionListItemSchemaMock()]
+      );
+      expect(esQuery).toEqual({
+        bool: {
+          filter: [
+            {
+              bool: {
+                filter: [
+                  {
+                    bool: {
+                      minimum_should_match: 1,
+                      should: [
+                        {
+                          match: {
+                            'host.name': 'linux',
+                          },
+                        },
+                      ],
+                    },
+                  },
+                  {
+                    bool: {
+                      filter: [
+                        {
+                          nested: {
+                            path: 'some.parentField',
+                            query: {
+                              bool: {
+                                minimum_should_match: 1,
+                                should: [
+                                  {
+                                    match: {
+                                      'some.parentField.nested.field': 'some value',
+                                    },
+                                  },
+                                ],
+                              },
+                            },
+                            score_mode: 'none',
+                          },
+                        },
+                        {
+                          bool: {
+                            must_not: {
+                              bool: {
+                                minimum_should_match: 1,
+                                should: [
+                                  {
+                                    match: {
+                                      'some.not.nested.field': 'some value',
+                                    },
+                                  },
+                                ],
+                              },
+                            },
+                          },
+                        },
+                      ],
+                    },
+                  },
+                ],
+              },
+            },
+          ],
+          must: [],
+          must_not: [],
+          should: [],
+        },
+      });
+    });
+
+    test('it should work with an empty list', () => {
+      const esQuery = getQueryFilter('host.name: linux', 'kuery', [], ['auditbeat-*'], []);
+      expect(esQuery).toEqual({
+        bool: {
+          filter: [
+            { bool: { minimum_should_match: 1, should: [{ match: { 'host.name': 'linux' } }] } },
+          ],
+          must: [],
+          must_not: [],
+          should: [],
+        },
+      });
+    });
+
+    test('it should work when lists has value undefined', () => {
+      const esQuery = getQueryFilter('host.name: linux', 'kuery', [], ['auditbeat-*'], []);
+      expect(esQuery).toEqual({
+        bool: {
+          filter: [
+            { bool: { minimum_should_match: 1, should: [{ match: { 'host.name': 'linux' } }] } },
+          ],
+          must: [],
+          must_not: [],
+          should: [],
+        },
+      });
+    });
+
+    test('it should work with a nested object queries', () => {
+      const esQuery = getQueryFilter(
+        'category:{ name:Frank and trusted:true }',
+        'kuery',
+        [],
+        ['auditbeat-*'],
+        []
+      );
+      expect(esQuery).toEqual({
+        bool: {
+          must: [],
+          filter: [
+            {
+              nested: {
+                path: 'category',
+                query: {
+                  bool: {
+                    filter: [
+                      {
+                        bool: {
+                          should: [
+                            {
+                              match: {
+                                'category.name': 'Frank',
+                              },
+                            },
+                          ],
+                          minimum_should_match: 1,
+                        },
+                      },
+                      {
+                        bool: {
+                          should: [
+                            {
+                              match: {
+                                'category.trusted': true,
+                              },
+                            },
+                          ],
+                          minimum_should_match: 1,
+                        },
+                      },
+                    ],
+                  },
+                },
+                score_mode: 'none',
+              },
+            },
+          ],
+          should: [],
+          must_not: [],
+        },
+      });
+    });
+
+    test('it works with references and does not add indexes', () => {
+      const esQuery = getQueryFilter(
+        '(event.module:suricata and event.kind:alert) and suricata.eve.alert.signature_id: (2610182 or 2610183 or 2610184 or 2610185 or 2610186 or 2610187)',
+        'kuery',
+        [],
+        ['my custom index'],
+        []
+      );
+      expect(esQuery).toEqual({
+        bool: {
+          must: [],
+          filter: [
+            {
+              bool: {
+                filter: [
+                  {
+                    bool: {
+                      filter: [
+                        {
+                          bool: {
+                            should: [{ match: { 'event.module': 'suricata' } }],
+                            minimum_should_match: 1,
+                          },
+                        },
+                        {
+                          bool: {
+                            should: [{ match: { 'event.kind': 'alert' } }],
+                            minimum_should_match: 1,
+                          },
+                        },
+                      ],
+                    },
+                  },
+                  {
+                    bool: {
+                      should: [
+                        {
+                          bool: {
+                            should: [{ match: { 'suricata.eve.alert.signature_id': 2610182 } }],
+                            minimum_should_match: 1,
+                          },
+                        },
+                        {
+                          bool: {
+                            should: [
+                              {
+                                bool: {
+                                  should: [
+                                    { match: { 'suricata.eve.alert.signature_id': 2610183 } },
+                                  ],
+                                  minimum_should_match: 1,
+                                },
+                              },
+                              {
+                                bool: {
+                                  should: [
+                                    {
+                                      bool: {
+                                        should: [
+                                          { match: { 'suricata.eve.alert.signature_id': 2610184 } },
+                                        ],
+                                        minimum_should_match: 1,
+                                      },
+                                    },
+                                    {
+                                      bool: {
+                                        should: [
+                                          {
+                                            bool: {
+                                              should: [
+                                                {
+                                                  match: {
+                                                    'suricata.eve.alert.signature_id': 2610185,
+                                                  },
+                                                },
+                                              ],
+                                              minimum_should_match: 1,
+                                            },
+                                          },
+                                          {
+                                            bool: {
+                                              should: [
+                                                {
+                                                  bool: {
+                                                    should: [
+                                                      {
+                                                        match: {
+                                                          'suricata.eve.alert.signature_id': 2610186,
+                                                        },
+                                                      },
+                                                    ],
+                                                    minimum_should_match: 1,
+                                                  },
+                                                },
+                                                {
+                                                  bool: {
+                                                    should: [
+                                                      {
+                                                        match: {
+                                                          'suricata.eve.alert.signature_id': 2610187,
+                                                        },
+                                                      },
+                                                    ],
+                                                    minimum_should_match: 1,
+                                                  },
+                                                },
+                                              ],
+                                              minimum_should_match: 1,
+                                            },
+                                          },
+                                        ],
+                                        minimum_should_match: 1,
+                                      },
+                                    },
+                                  ],
+                                  minimum_should_match: 1,
+                                },
+                              },
+                            ],
+                            minimum_should_match: 1,
+                          },
+                        },
+                      ],
+                      minimum_should_match: 1,
+                    },
+                  },
+                ],
+              },
+            },
+          ],
+          should: [],
+          must_not: [],
+        },
+      });
+    });
+  });
+});
